refactor(artifacts): Artifacts: consolidate manifest entry classes (#5547)
